Split SiloHostBuilder extensions in two separate files
